+++ lib/utils/date.h
@@ -5,7 +5,7 @@
 class DateUtils
 {
 public:
-    static unsigned int now()
+    static uint64_t now()
     {
         return std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count();
     }
